requirements.txt
# MEAL requirements
# Usage: pip install -r requirements.txt

albumentations==1.3.0
comet_ml==3.33.3
coremltools==6.3.0
einops==0.6.1
Flask==2.3.2
GitPython==3.1.31
ipython==8.11.0
matplotlib==3.4.3
mss==9.0.1
natsort==8.3.1
numpy==1.23.4
onnx==1.14.0
onnxruntime==1.15.0
onnxsim==0.4.28
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.68
openvino==2022.3.0
#paddle==1.0.2
pafy==0.5.5
pandas==1.5.3
Pillow==9.5.0
psutil==5.9.4
pycocotools==2.0.6
PyYAML==6.0
Requests==2.31.0
scikit_learn==1.2.1
scipy==1.8.1
seaborn==0.12.2
setuptools==65.5.0
tensorflow==2.12.0
tensorflowjs==4.6.0
tensorrt==8.6.1
tensorboard>=2.4.1
tflite_runtime==2.12.0
tflite_support==0.4.3
thop==0.1.1.post2209072238
timm==0.6.12
#torch==1.13.0+cu116
torchsummary==1.5.1
#torchvision==0.14.0+cu116
tqdm==4.64.1
tritonclient==2.33.0
webcolors==1.12
x2paddle==1.4.1
